% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvolutePixels.R
\name{ConvolutePixels}
\alias{ConvolutePixels}
\title{ConvolutePixels}
\usage{
ConvolutePixels(image, convolution.factor = 2)
}
\arguments{
\item{image}{rgbimage matrix object.}

\item{convolution.factor}{n x n area of pixels to reduce.}
}
\value{
rgbimage
}
\description{
Takes a bitmap image in the form of a matrix and convolutes the pixels to form
an image of a smaller size.
}
\examples{

image = ImportPicture(system.file("pictures", "pngexample.png", package="PictureToMinecraft"))
dim(image)
image.conv = ConvolutePixels(image, convolution.factor = 20)
dim(image.conv)

}
