% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawBlock.R
\name{DrawBlock}
\alias{DrawBlock}
\title{DrawBlock}
\usage{
DrawBlock(filename, convolute.blocks.factor = 1)
}
\arguments{
\item{filename}{filename for image}
}
\value{
plot
}
\description{
Draws a Block face given a minecraft image
}
\examples{

image.filename = 'grass_side_carried.png'
DrawBlock(image.filename)

}
