% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportPicture.R
\name{ImportPicture}
\alias{ImportPicture}
\title{ImportPicture}
\usage{
ImportPicture(file)
}
\arguments{
\item{file}{file path with respect to working directory.}
}
\value{
gsimage
}
\description{
Imports a .png file from a file location and converts it to grey scale.
Uses the *png* library.
}
\examples{

file = system.file("pictures", "pngexample.png", package="PictureToMinecraft")
image = ImportPicture(file)
class(image)
dim(image)



}
