% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PixelToBlock.R
\name{PixelToBlock}
\alias{PixelToBlock}
\title{PixelToBlock}
\usage{
PixelToBlock(x)
}
\arguments{
\item{x}{pixel vector of size 3.}
}
\value{
character vector corresponding to the block filename.
}
\description{
PixelToBlock
}
\examples{

green = c(0, 1, 0)
PixelToBlock(green)

}
