% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotImage.R
\name{PlotImage}
\alias{PlotImage}
\title{PlotImage}
\usage{
PlotImage(image)
}
\arguments{
\item{image}{rgbimage}
}
\value{
plot
}
\description{
Plots an RGB image (rgbimage)
}
\examples{

image = ImportPicture(system.file("pictures", "pngexample.png", package="PictureToMinecraft"))
PlotImage(image)

}
