% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMinecraftMap.R
\name{PlotMinecraftMap}
\alias{PlotMinecraftMap}
\title{PlotMinecraftMap}
\usage{
PlotMinecraftMap(image, convolute.blocks.factor = 2)
}
\arguments{
\item{image}{rgbimage object}
}
\value{
plot
}
\description{
Takes a rgbimage matrix image bitmap image object and returns a dice map plot.
}
\examples{

image = ImportPicture(system.file("pictures", "pngexample.png", package="PictureToMinecraft"))
dim(image)
image.conv = ConvolutePixels(image, convolution.factor = 100)
dim(image.conv)
image.prepped = PrepareMinecraftMap(image.conv)
PlotMinecraftMap(image.prepped)


}
