% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareMinecraftMap.R
\name{PrepareMinecraftMap}
\alias{PrepareMinecraftMap}
\title{PrepareMinecraftMap}
\usage{
PrepareMinecraftMap(image)
}
\arguments{
\item{image}{rgbimage}
}
\value{
rgbimage.minecraft
}
\description{
PrepareMinecraftMap
}
\examples{

image = ImportPicture(system.file("pictures", "pngexample.png", package="PictureToMinecraft"))
image[,,1][1:5]
image.conv = ConvolutePixels(image, convolution.factor = 200)
image.prepped = PrepareMinecraftMap(image.conv)
image[1:5]


}
