% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToHex.R
\name{ToHex}
\alias{ToHex}
\title{ToHex}
\usage{
ToHex(x)
}
\arguments{
\item{x}{vector}
}
\value{
character hexcode
}
\description{
Take a 3D vector RGB and converyt to a single Hexidecimal string.
}
\examples{

x = 1:3 * 0.1
ToHex(x)

}
